requirements.txt
Flask==2.0.1
transformers==4.15.0
PyPDF2==1.26.0
python-docx==0.8.11
